cpp-port: batch 83 - port 22 hw/ files, fix headers, improve batch-port.py
Port 22 hw/ files from C to C++:
- hw/9pfs: cofile, cofs
- hw/char: debugcon
- hw/display: edid-region
- hw/i386: monitor, nitro_enclave, kvm/i8259
- hw/misc: debugexit, pc-testdev, unimp
- hw/net: rocker/rocker_world
- hw/virtio: vhost-iova-tree, vhost-scsi-pci, vhost-user-blk-pci,
  vhost-user-scsi-pci, virtio-9p-pci, virtio-balloon-pci,
  virtio-blk-pci, virtio-iommu-pci, virtio-net-pci, virtio-scsi-pci,
  virtio-serial-pci

Header fixes for C++ compatibility:
- Add extern "C" guards: iova-tree.h, vhost-iova-tree.h, i8259.h,
  i8259_internal.h, mc146818rtc.h, eif.h, 9p.h, file-op-9p.h,
  virtio-scsi.h (include ordering fix)
- DEV_NVECTORS_UNSPECIFIED: constexpr uint32_t in C++ to avoid
  sign-compare warnings
- target/arm/cpu.h: rename 'new' parameter to 'new_val'
- Clean up redundant static_cast<uint32_t>(DEV_NVECTORS_UNSPECIFIED)
  in existing .cpp files

C++ fixes in ported files:
- void* casts (static_cast) for opaque/gpointer/g_malloc
- Designator initializer reordering (MemoryRegionOps, TypeInfo)
- Sign-compare fixes (int vs uint32_t loop vars)

Improve batch-port.py:
- Better linker error detection (map .o files to .cpp sources)
- Add BLOCKED_DIRS for hw/arm, hw/ppc, hw/riscv (target cpu.h)
- Save build output to /tmp on unidentifiable failures

+++ hw/virtio/vhost-user-fs-pci.cpp
@@ -39,7 +39,7 @@ static void vhost_user_fs_pci_realize(VirtIOPCIProxy *vpci_dev, Error **errp)
     VHostUserFSPCI *dev = VHOST_USER_FS_PCI(vpci_dev);
     DeviceState *vdev = DEVICE(&dev->vdev);
 
-    if (vpci_dev->nvectors == static_cast<uint32_t>(DEV_NVECTORS_UNSPECIFIED)) {
+    if (vpci_dev->nvectors == DEV_NVECTORS_UNSPECIFIED) {
         /* Also reserve config change and hiprio queue vectors */
         vpci_dev->nvectors = dev->vdev.conf.num_request_queues + 2;
     }
